project-root/requirements.txt
# Web framework
fastapi
uvicorn[standard]

# GitHub App
PyJWT>=2.8.0
cryptography
httpx
PyGithub

# Parsing
pyyaml

# AI
google-generativeai

# Env
python-dotenv

# CORS
starlette